[refactor] #236 위니 스낵바 클래스를 추상 클래스로 변경
공통적인 내용은 WineySnackbar 클래스에서 상속 받고, 스낵바의 각 타입에 따라 자식 클래스에서 오버라이딩 하는 함수 내용이 다르도록 변경
