+++ app/src/main/java/org/go/sopt/winey/util/context/ContextExt.kt
@@ -2,6 +2,7 @@ package org.go.sopt.winey.util.context
 
 import android.app.Activity
 import android.content.Context
+import android.view.Gravity
 import android.view.View
 import android.view.inputmethod.InputMethodManager
 import android.widget.Toast
@@ -11,8 +12,10 @@ import androidx.annotation.StringRes
 import androidx.core.content.ContextCompat
 import com.google.android.material.snackbar.Snackbar
 import org.go.sopt.winey.R
-import org.go.sopt.winey.util.view.SnackbarType
-import org.go.sopt.winey.util.view.WineySnackbar
+import org.go.sopt.winey.util.view.snackbar.NotiPermissionSnackbar
+import org.go.sopt.winey.util.view.snackbar.SnackbarType
+import org.go.sopt.winey.util.view.snackbar.WineyFeedResultSnackbar
+import org.go.sopt.winey.util.view.snackbar.WineySnackbar
 
 /** Hide keyboard from window */
 fun Context.hideKeyboard(view: View) {
@@ -37,14 +40,32 @@ fun Context.wineySnackbar(
     message: String,
     type: SnackbarType
 ) {
-    val snackbar = WineySnackbar.make(anchorView, message, type)
-    if (type is SnackbarType.NotiPermission) {
-        snackbar.setAction(
-            resId = R.string.snackbar_noti_permission_setting_text,
-            onClicked = type.onActionClicked
-        )
+    when (type) {
+        is SnackbarType.WineyFeedResult -> {
+            WineyFeedResultSnackbar(
+                anchorView = anchorView,
+                gravity = Gravity.TOP,
+                message = message
+            ).apply {
+                initResultIcon(isSuccess = type.isSuccess)
+                show()
+            }
+        }
+
+        is SnackbarType.NotiPermission -> {
+            NotiPermissionSnackbar(
+                anchorView = anchorView,
+                gravity = Gravity.TOP,
+                message = message
+            ).apply {
+                setAction(
+                    resId = R.string.snackbar_noti_permission_setting_text,
+                    onClicked = type.onActionClicked
+                )
+                show()
+            }
+        }
     }
-    snackbar.show()
 }
 
 fun Context.stringOf(@StringRes resId: Int) = getString(resId)
